Revert "fix: use correct CMakeLists.txt file for Android"
This reverts commit eef3b55838a40d7e827b150b64c9ddc51af54844.

In the release package there is no symlink, but the folder structure is
different. This means with this change the release can't be built
anymore. This change was primarily meant to fix the git build on Windows
systems without symlink support. The proper solution for that will be to
get rid of the symlinks by integrating the upstream olm as a submodule.
To fix the release revert this chnange for now.